[droidian-release] snapshot-control: add and ship an i386-only reference to the current Debian snapshot
Droidian snapshots don't export i386 packages as they're not needed.
However, on the buildd side there might be some need to have i386 packages
around. This support file supplies the i386 repository for the same Debian
snapshot used as a base.

This is not enabled by default, and you should symlink this in
/etc/apt/sources.list.d.

This is part of a package (rather than be a buildd-side quirk) so that it
can be easily updated (especially useful if using the flavour symlink
rather than the static snapshot version).

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>