[droidian-snapshot] Move feature branch handling in droidian-snapshot
This removes the need for the feature branch boilerplate in the
control file.

Currently just the standard repos + droidian-gst-plugins-bad1.0 + droidian-release
are kept in the control file. Everything else has been moved (beyond some
outdated branches that aren't needed anymore in Droidian 100).

This will allow us to move to a more modular buildd workflow.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>